Complete the landing page
This took me two days to do but I was able to make it look a lot like a
landing page. Afterwards i cleaned up my code as well so I'm happy with
the outcome. I should have commited more (or at all) but i was working
through it. I really don't understand the flex command and will probably
watch a video on it